+++ lib/TItemContainer.php
@@ -1,4 +1,19 @@
 <?php
+/*
+   +----------------------------------------------------------------------+
+   | LiteRT Delay Initializer Library                                     |
+   +----------------------------------------------------------------------+
+   | Copyright (c) 2007-2017 Fenying Studio                               |
+   +----------------------------------------------------------------------+
+   | This source file is subject to version 2.0 of the Apache license,    |
+   | that is bundled with this package in the file LICENSE, and is        |
+   | available through the world-wide-web at the following url:           |
+   | https://github.com/litert/delay-initializer/blob/master/LICENSE      |
+   +----------------------------------------------------------------------+
+   | Authors: Angus Fenying <i.am.x.fenying@gmail.com>                    |
+   +----------------------------------------------------------------------+
+ */
+
 declare (strict_types = 1);
 
 namespace L\Kits\DelayInit;
@@ -7,9 +22,9 @@ use \L\Core\Exception;
 
 /**
  * This trait provides the basic methods for managing the delay
- * initliazed items.
+ * initialized items.
  *
- * @package L\Kits\DelayInit
+ * @package litert/delay-initializer
  */
 trait TItemContainer
 {
@@ -54,7 +69,8 @@ trait TItemContainer
         if (!is_callable($this->__delayCtors[$name] ?? false)) {
 
             throw new Exception(
-                "Item {$name} not found."
+                "Item {$name} not found.",
+                \L\Error\DelayInit\ITEM_NOT_FOUND
             );
         }
 
@@ -69,7 +85,10 @@ trait TItemContainer
      * @param string $name           Name of item.
      * @param callable $initializer  Initializer of item.
      */
-    public function setInitializer(string $name, callable $initializer)
+    public function setInitializer(
+        string $name,
+        callable $initializer
+    )
     {
         $this->__delayCtors[$name] = $initializer;
     }