+++ lib/TPropertyContainer.php
@@ -1,4 +1,19 @@
 <?php
+/*
+   +----------------------------------------------------------------------+
+   | LiteRT Delay Initializer Library                                     |
+   +----------------------------------------------------------------------+
+   | Copyright (c) 2007-2017 Fenying Studio                               |
+   +----------------------------------------------------------------------+
+   | This source file is subject to version 2.0 of the Apache license,    |
+   | that is bundled with this package in the file LICENSE, and is        |
+   | available through the world-wide-web at the following url:           |
+   | https://github.com/litert/delay-initializer/blob/master/LICENSE      |
+   +----------------------------------------------------------------------+
+   | Authors: Angus Fenying <i.am.x.fenying@gmail.com>                    |
+   +----------------------------------------------------------------------+
+ */
+
 declare (strict_types = 1);
 
 namespace L\Kits\DelayInit;
@@ -7,9 +22,9 @@ use \L\Core\Exception;
 
 /**
  * This trait provides the basic methods for managing the delay
- * initliazed properties.
+ * initialized properties.
  *
- * @package L\Kits\DelayInit
+ * @package litert/delay-initializer
  */
 trait TPropertyContainer
 {
@@ -38,7 +53,9 @@ trait TPropertyContainer
      * Get a delay-initialized property by name.
      *
      * @param string $name  name of property
+     *
      * @return mixed
+     *
      * @throws Exception
      */
     public function __get(string $name)
@@ -46,7 +63,8 @@ trait TPropertyContainer
         if (!is_callable($this->__delayCtors[$name] ?? false)) {
 
             throw new Exception(
-                "Property {$name} not found."
+                "Property {$name} not found.",
+                \L\Error\DelayInit\ITEM_NOT_FOUND
             );
         }
 
@@ -60,6 +78,8 @@ trait TPropertyContainer
      *
      * @param string $name           Name of property.
      * @param callable $initializer  Initializer of property.
+     *
+     * @return void
      */
     public function setInitializer(
         string $name,