+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changed Logs
 
+## v0.2.0 (2017-10-01)
+
+- Added interface `ItemContainerEx`.
+- Added interface `PropertyContainerEx`.
+- Improved the comments for methods and interfaces.
+- Added error codes.
+
 ## v0.1.2 (2017-08-24)
 
 - Add method `_setInitializerArgs` to setup the arguments to the initializers.